Validate tip header merkle roots during horizon sync
This PR creates MMR checkpoints for each "chunk" of downloaded
kernels/UTXOs. This commits the in-memory checkpoint to persistence
thereby reducing the memory footprint during syncing. This will allow
hundreds of thousands of hashes to be downloaded without memory bloat
and allows the merkle root state to be validated after each sync step
rather than at the end of the sync.

Once the horizon state has been synchronized, the checkpoints are merged
into a single checkpoint as part of `commit_horizon_state`.

If any part of the sync fails, the database will be in an inconsistent
state. A subsequent PR will aim to address this by storing metadata
about the last validated state, and using that to return the database to
that state when entering HorizonSync mode.

- Create MMR checkpoints for each "chunk" of requests (Kernel/UTXO).
- Validate merkle roots against the horizon header after kernel and UTXO
  sync stages. To facilitate this, `validate_mmr_root` was added to the
  `BlockchainBackend` trait.
- `validate_mmr_root` checks the tip header merkle roots against the
  current checkpoint state.
- Use a single retry config option for all horizon sync requests. There
  should be no need to specify a different number of request retries for
  each type of request.
- Retry each sync stage up to the number of sync peers available.
  Previously, this was a configurable number. Since a peer is
  banned/removed if anything goes wrong at a particular stage, it does
  not make sense to have a retry configuration greater than the number
  of sync peer available.
- Made the `MmrTree` enum a `Copy` type since its underlying representation is
  trivially small. Removed all explicit clones to keep clippy happy.
- Various logging improvements